Initial ViaQ documentation site.
From the initial page:

The ViaQ project provides components for  https://docs.openshift.com/container-platform/4.7/logging/cluster-logging.html[Openshift Logging], a central log collection system for openshift clusters.

This site documents the architecture and some of the internal details of Openshift Logging.
It is aimed at developers, doc writers, and contributors working on the project,
but may be of interest to those who want to learn more about the internal details.

Actions taken so far to build the site:
- Move architecture documentation from the CLO repo.
- Extract data model from [[https://github.com/ViaQ/elasticsearch-templates][viaQ/elasticsearch-templates repo]].
  - Removed unused namespaces and Elasticsearch-specific generation tools.
  - Consistent file extensions: .asciidoc -> .adoc, .yml -> .yaml
  - Simplified YAML model.
  - Simple templates generate asciidoc snippets for inclusion in larger doc.

Needs full review to ensure nothing important has been dropped.

All docs need to be reviewed and brought up to date.

There are many gaps and omissions.